fix: portable c_char, mode_t, time_t, and errno handling across targets
Bionic's c_char is u8 and 32-bit targets have narrower time_t/c_long,
so hardcoded i8/i64 casts broke cross-compilation (reported for
aarch64-linux-android / armv7-linux-androideabi; also confirmed on the
release-matrix musl ARM targets):

- Replace *const/*mut i8 casts and annotations with libc::c_char
  throughout src/ and crates/abp (argv/envp, sun_path, ifr_name).
- Width-safe `as u32` casts on st_mode & S_IFMT comparisons; new
  io::is_dir()/is_reg() helpers in abp replace duplicated cast chains.
- Explicit libc::time_t / c_long casts in 13 applets that assumed
  64-bit time_t (touch, w, who, sleep, dmesg, rtcwake, linux32,
  iorenice, getconf, ts, traceroute, arping, sntp).
- Use Bionic's __errno() (not __errno_location()) on Android in
  sys::errno()/clear_errno() and the new abp io::errno()/set_errno();
  route five applets' raw __errno_location() reads through sys::errno().
- abp io path helpers now set ENAMETOOLONG on the too-long-path short
  circuit so mkdir_tracked() can't mistake stale EEXIST for success.

Verified: 0 errors on aarch64-unknown-linux-musl and
armv7-unknown-linux-musleabihf (release profile), abp checks clean on
both Android targets, host release build green, 1468 tests pass.

Reported-by: GitHub issue on abp Android cross-compilation

+++ src/applets/file/ln.rs
@@ -70,7 +70,7 @@ pub fn ln(argc: i32, argv: *const *const u8) -> i32 {
     // of being dereferenced.
     let mut st: libc::stat = unsafe { core::mem::zeroed() };
     let stat_ret = if no_deref { io::lstat(last, &mut st) } else { io::stat(last, &mut st) };
-    let last_is_dir = stat_ret == 0 && (st.st_mode & libc::S_IFMT) == libc::S_IFDIR;
+    let last_is_dir = stat_ret == 0 && (st.st_mode as u32 & libc::S_IFMT as u32) == libc::S_IFDIR as u32;
 
     if last_is_dir {
         // ln [-fs] source... target_dir
